# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Sprout API
  version: 0.1.0
  description: API for operating and sharing small applications.
servers:
  - url: http://127.0.0.1:8080
    description: Local development
tags:
  - name: Applications
  - name: Identity
paths:
  /api/v1/me:
    get:
      tags: [Identity]
      summary: Resolve the signed-in user and home workspace
      security:
        - clerkSession: []
      responses:
        "200":
          description: Signed-in Sprout identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: Identity provider or database unavailable
  /api/v1/workspaces/{workspaceSlug}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceSlug"
    get:
      tags: [Identity]
      summary: Get a workspace through verified membership
      security:
        - clerkSession: []
      responses:
        "200":
          description: Workspace available to the member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/workspaces/{workspaceSlug}/applications:
    parameters:
      - $ref: "#/components/parameters/WorkspaceSlug"
    get:
      tags: [Applications]
      summary: List applications available to the workspace member
      security:
        - clerkSession: []
      responses:
        "200":
          description: Applications in most-recently-updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Applications]
      summary: Create an application in the workspace
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplication"
      responses:
        "201":
          description: Application created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/workspaces/{workspaceSlug}/applications/{applicationId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceSlug"
      - $ref: "#/components/parameters/ApplicationID"
    get:
      tags: [Applications]
      summary: Get one application available to the workspace member
      security:
        - clerkSession: []
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Applications]
      summary: Update application details or lifecycle
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplication"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session token verified by the Go API and mapped to a Sprout user.
  parameters:
    WorkspaceSlug:
      name: workspaceSlug
      in: path
      required: true
      schema:
        type: string
        maxLength: 63
        pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    ApplicationID:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Workspace:
      type: object
      additionalProperties: false
      required: [id, slug, name]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
    Me:
      type: object
      additionalProperties: false
      required: [id, email, displayName, workspace]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        workspace:
          $ref: "#/components/schemas/Workspace"
    ApplicationLifecycle:
      type: string
      enum: [active, paused, archived]
    Application:
      type: object
      additionalProperties: false
      required: [id, name, slug, description, lifecycle, defaultHostname, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        lifecycle:
          $ref: "#/components/schemas/ApplicationLifecycle"
        defaultHostname:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApplicationList:
      type: object
      additionalProperties: false
      required: [applications]
      properties:
        applications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
    CreateApplication:
      type: object
      additionalProperties: false
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        slug:
          type: string
          maxLength: 63
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        description:
          type: string
          maxLength: 2000
          default: ""
    UpdateApplication:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 2000
        lifecycle:
          $ref: "#/components/schemas/ApplicationLifecycle"
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
  responses:
    BadRequest:
      description: Invalid path, body, or lifecycle transition
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid development identity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Workspace role does not permit this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Workspace or application not available to this identity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Application slug already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Unexpected internal failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"